Starting to crawl tweets...
Some files drawing the structure of the crawler.

+ crawler/HttpUtils.py includes a simple function to do a HTTP request
(GET or POST). We should include some timeout, because proxies may be
extremely slow!
Any function or class regarding the HTTP protocol will be placed here,
if it is not too big/complex.

+ crawler/OsUtils.py includes a function called 'safemkdir' that
creates a directory (and its parents if they do no exist) if it did not
exist. This function masks all the exceptions, so it just returns True
if the directory was successfully created or it already exist,
otherwise it return False.

+ crawler/ProxiedRequester.py includes the class ProxiedRequester which
will try to send a HTTP request to several proxy servers.

+ crawler/TwitterCrawler.py includes the class TwitterCrawler. This
will start from a set of twitter users and will get all of their
tweets, retweets, replies and mentions. Then, the users referenced will
be added to the frontier and the process will be repeated. It is very
important that the crawler minimizes the number of queries! So, we will
store on disk any 'tweet' (status) or user data fetched from Twitter.
The id of the tweet or the user will be added to a set, so we can avoid
doing repeated queries.

+ data/ will contain two subfolders 'tweets' and 'users', where the
downloaded tweets and user details will be stored.