Mathematical approach to TweetRank
In this document, I expressed some of the ideas that we discussed in
the first meeting and some others that I think that can be interesting
to solve our problem.

We should review this document, and send a copy to the proposer of the
project (and maybe Hedvig too) to let him know what are we thinking to
do.

The content is quite mathematical, but I think that it is important to
express properly what we want to do. Try to read it and we can discuss
it on the next meeting.