Working on Crawler
The crawler is able to fetch all the tweets (including retweets and
replies) starting from a set of users, and fetch recursively tweets
from the users mentioned in the tweets.

There are many things to correct and improve, but this is enough to get
a reasonable amount of tweets to start working.