Crawler improved, tweets storage changed
Important update: Now tweets are not stored in separated files, but in
a single file for each user. TweetWarehouse is not needed anymore.
UsersWarehous.py is the one that stores the tweets on disk through the
method add_tweets_to_user().

Minor fixes on HttpUtils.py and ProxiedRequester.py.

The module mysolr-custom is a customized version of mysolr-0.5.1
(http://mysolr.redtuna.org/). The original module had problems with
python3 and unicode characters when parsing JSON responses. This issue
has been corrected.

TODO: Insert tweets to Solr and insert the relationships needed to
compute the TweetRank to the Ranker.