Mathematical definition updated
I have updated the mathematical description of our approximation to
include the hashtag relation among tweets (this is not the same that
the similarity weighting between users with common hashtags).

I have also noticed that implementing the similarity weighting may be
trickier than I thought the first time. At least, the naïve solution
would be very costly. I think that I have found an equivalent way to do
some computations but I am not sure if it is really equivalent and to
prove formally that is quite a mess! So, we should discuss if we will
finally implement the probability weighting.