Graph on disk...
We should store the graph used to compute the PageRank on disk.
MegaGraph is a class that does it and it is used by the Ranker and the
HTTPServ modules.

Also, old XMLParser.java and related files deleted.

By the way, I added a couple of files for the Eclipse configuration.