Ranker working (almost completed!)
I didn't manage to work with MegaMap, so I used the Hashtable classes
again to represent the graph, but now, I added a method to store these
objects on disk, so we can recover them if anything goes wrong.

I did not removed the MegaGraph from the repository because we should
use it.