Performance improved for Graph
The old class Graph has been divided into two different classes:
PersistentGraph and TemporaryGraph.

PersistentGraph is used by httpserv module classes to add the new
information provided by the crawler. The internal structure for
adjacency lists is HashSet to avoid duplicated data efficiently.

TemporaryGraph is used by TweetRanker. This graph is generated as a
copy of PersistentGraph (we need a copy to ensure that no new data is
added while computing the TweetRank), the difference is that the
internal structure for adjacency lists is an ArrayList to allow random
access, which improves the performance of TweetRanker.

Some Condition Variables should be used to disallow the modification of
the PersistentGraph while it is being copied.