TweetRank working!
Many changes made on ranker package!

- Names of packages and classes have changed. For example, the old
RankerDataServer is now called TweetRanker and it has been moved to the
"default" package. The old TweetRanker class is called
TweetRankComputer now, inside the computer package (the only purpose of
this class is to compute the PageRank).

- log4j is used by all classes to handle the log.

- A shutdown hook has been added to TweetRanker so, when the
application is terminated, the Persistent Graph is stored on disk.

- Periodically, the Persistent graph is stored on disk (every 20
minutes) and the TweetRank is refreshed (every 30 minutes).

- The computation of the TweetRank has been protected with a Lock to
avoid concurrent computations. Concurrent computations do not cause any
trouble, but they use many resources of the computer.

TODO: Parallelize the computation of TweetRank in multiple threads.

TODO2: Test the ranking!