TweetRankComputer corrected
- The Monte Carlo method was not correctly implemented. Due the correct
implementation of the method, which has a non-deterministic running
time, the methods for getting the percentage of completion and the
estimated remaining time have been removed.
A new concept of BORED_PROBABILITY has been introduced. At the end of
each jump, if rand() > BORED_PROBABILITY, then the random walk finishes.

- There was a bug assigning the new currentID. Basically, it overwrote
the currentID variable with null.

- Status Handler modified to show the new status view.