Many changes involving notification
DirectoryNotification: It was extremely slow and extremely memory
consuming. SAX parsing is now used instead of DOM. Much more fast and
memory efficient!

The HTTP Server in the ranker was not able to handle concurrent
connection. A new library developed by Ziyan Zhou is used, the
RequestHandler now handles both DataRequests and StatusRequests.

TODO: The crawler should be adapted to the new SAX parsing!