RankerNotifier and Handler modified to allow multiple line requests.
With this modification, a single HTTP requests can encode multiple
operations on the Ranker graph. This improvement reduces the number of
HTTP requests needed to update the Ranker and improves the performance.

The probabilities parameters of the TweetRank algorithm have changed,
the previous ones were too much equally distributed, but we still need
to do some experiments on these parameters.