TweetRanker improved, now stops at dangling nodes
Major changes:

When a TweetRank computation starts, it makes a copy of the current
PersistentGraph, instead of loading the latest file on disk. Fresher
information & faster load.

TweetRank computation is performed using the Monte Carlo Comple Path
method Stopping at Dangling Nodes. This method is faster and provides a
better approximation of the TweetRank (look at paper: "Monte Carlo
methods in PageRank computation: When one iteration is sufficient").

The approximated percentage of completion time and the approximated
remaining time for the TweetRank computation is calculated to be shown
in the status.

TODO:
- Use a ThreadPool instead of creating new threads each computation
starts.
- Run experiments, to see how the algorithm scales in the number of
tweets.
- It would be interesting to run experiments to find the optimal
parameters of the algorithm, but this is hard to do, if we can't
compare our search results.